修改密码

+++ sky-server/src/main/java/com/sky/mapper/EmployeeMapper.java
@@ -48,5 +48,5 @@ public interface EmployeeMapper {
      * @return
      */
     @Select("select * from employee where id=#{id};")
-    Employee selectById(Integer id);
+    Employee selectById(Long id);
 }
